Move URL placeholder code into Utilities.swift
