+++ Plash/Utilities.swift
@@ -2603,6 +2603,42 @@ extension URL {
 }
 
 
+extension URL {
+	enum PlaceholderError: Error, LocalizedError {
+		case failedToEncodeToken(String)
+		case invalidURLAfterSubstitution(String)
+
+		var errorDescription: String? {
+			switch self {
+			case .failedToEncodeToken(let token):
+				return "Failed to encode token “\(token)”"
+			case .invalidURLAfterSubstitution(let urlString):
+				return "New URL was not valid after substituting placeholders. URL string is “\(urlString)”"
+			}
+		}
+	}
+
+	/**
+	Replaces any occurences of `placeholder` in the url with `replacement`.
+	Throws an error if the replacement would create an invalid URL.
+	*/
+	func replacingPlaceholder(_ placeholder: String, with replacement: String) throws -> URL {
+		guard let token = placeholder.addingPercentEncoding(withAllowedCharacters: .urlPathAllowed) else {
+			throw PlaceholderError.failedToEncodeToken(placeholder)
+		}
+
+		let urlString = absoluteString
+			.replacingOccurrences(of: token, with: replacement)
+
+		guard let newURL = URL(string: urlString) else {
+			throw PlaceholderError.invalidURLAfterSubstitution(urlString)
+		}
+
+		return newURL
+	}
+}
+
+
 struct Reachability {
 	/// Checks whether we're currently online.
 	static func isOnline(host: String = "apple.com") -> Bool {
